+++ federated_gradient.py
@@ -1,7 +1,10 @@
 from gradient_descent import get_local_gradient
+from tools import gradient_aggregate
 
 number_locations = 5
+theta0_list = []
+theta1_list = []
 
 
 
-
+for location in range(number_locations):
